Refactor: 온라인 게임 설정 폴더 분리
