+++ FE/src/pages/online/gameSetting/OnlineGameSettingTournament.js
@@ -1,9 +1,9 @@
-import { changeUrl, changeUrlData } from '../../index.js';
-import HorizontalButton from '../../components/HorizontalButton.js';
-import VerticalButton from '../../components/VerticalButton.js';
-import ButtonBackArrow from '../../components/ButtonBackArrow.js';
-import { pongImage } from '../../components/pongImage.js';
-import { createRoomAPI } from './createRoomAPI.js';
+import { changeUrl, changeUrlData } from '../../../index.js';
+import HorizontalButton from '../../../components/HorizontalButton.js';
+import VerticalButton from '../../../components/VerticalButton.js';
+import ButtonBackArrow from '../../../components/ButtonBackArrow.js';
+import { pongImage } from '../../../components/pongImage.js';
+import { createAndJoinRoom } from '../rooms/roomManager.js';
 
 const html = String.raw;
 
@@ -172,7 +172,7 @@ class OnlineGameSettingTournament {
 			this.resetData();
 			updateRoomName(newData);
 			newData.total_score *= 5;
-			createRoomAPI(newData);
+			createAndJoinRoom(newData);
 		});
 
 		const backButton = document.querySelector('.button-back-in-window');
